requirements.txt
atlassian-python-api==4.0.4
beautifulsoup4==4.13.4
certifi==2025.7.14
charset-normalizer==3.4.2
Deprecated==1.2.18
idna==3.10
jmespath==1.0.1
markdown2==2.5.3
oauthlib==3.3.1
PyYAML==6.0.2
requests==2.32.4
requests-oauthlib==2.0.0
soupsieve==2.7
typing-extensions==4.14.1
urllib3==2.5.0
wrapt==1.17.2
